fixed scalings

+++ src/generators/einstein/cli.py
@@ -61,7 +61,7 @@ def render_pattern(
 def build_parser():
     parser = argparse.ArgumentParser(description="Generate aperiodic Einstein tile images.")
     parser.add_argument("--iterations", type=int, default=DEFAULT_ITERATIONS, help="Number of subdivision rounds to render.")
-    parser.add_argument("--scalar", type=int, default=DEFAULT_SCALAR, help="Pixel scale for each tile coordinate.")
+    parser.add_argument("--scalar", type=float, default=DEFAULT_SCALAR, help="Pixel scale for each tile coordinate.")
     parser.add_argument("--width", type=int, default=DEFAULT_WIDTH, help="Output image width in pixels.")
     parser.add_argument("--height", type=int, default=DEFAULT_HEIGHT, help="Output image height in pixels.")
     parser.add_argument("--center-x", type=float, default=0, help="Pattern-space X coordinate placed at the canvas center.")
